Fix slightly snide comment about SQLite
The original language implies that SQLite is not an "actual" database engine, which is of course not the case! SQLite is a very capable and ubiquitous database engine. It's not appropriate for many applications, though, so perhaps the document here can simply read "different" rather than "actual". I recognize this is a bit particular, but hey-- that's how I perceived the language on my first read, and I don't even particularly like (or use) SQLite anyways! I do respect the technology, though, and thought it'd be kind to change the language so as to remove all possibility of misinterpretation.